fix(Zip): repair fixes and indentation, add class ZipException

+++ src/Utility/Zip.php
@@ -2,6 +2,7 @@
 
 namespace Solcre\SolcreFramework2\Utility;
 
+use Solcre\SolcreFramework2\Exception\ZipException;
 use ZipArchive;
 
 class Zip
@@ -9,34 +10,44 @@ class Zip
     public static function compressAndSend($dir, array $files, $zipname): void
     {
         $file = tempnam("/tmp", "zip");
-        $zip = new ZipArchive();
-        $zip->open($file, \ZipArchive::OVERWRITE);
-        if (is_array($files) && count($files)) {
-            foreach ($files as $name) {
-                $zip->addFile($dir . $name, $name);
+
+        if ($file !== false) {
+            $zip = new ZipArchive();
+            $zip->open($file, \ZipArchive::OVERWRITE);
+
+            if (is_array($files) && count($files)) {
+                foreach ($files as $name) {
+                    $zip->addFile($dir . $name, $name);
+                }
             }
+
+            $zip->close();
+            header('Content-Type: application/zip');
+            header('Content-Length: ' . filesize($file));
+            header('Content-Disposition: attachment; filename="' . $zipname . '.zip"');
+            readfile($file);
+            unlink($file);
+            exit;
         }
-        $zip->close();
-        header('Content-Type: application/zip');
-        header('Content-Length: ' . filesize($file));
-        header('Content-Disposition: attachment; filename="' . $zipname . '.zip"');
-        readfile($file);
-        unlink($file);
-        exit;
+
+        throw ZipException::temporaryFilenameException();
     }
 
-    public static function extract($filename, $location = null, $delete = false)
+    public static function extract($filename, $location = null, $delete = false): bool
     {
-        $success = false;
-        $location = $location ? $location : dirname($filename);
-        $zip = new \ZipArchive();
+        $success  = false;
+        $location = $location ?: dirname($filename);
+        $zip      = new \ZipArchive();
+
         if ($zip->open($filename) === true) {
             $success = $zip->extractTo($location);
             $zip->close();
+
             if ($success && $delete) {
                 unlink($filename);
             }
         }
+
         return $success;
     }
 
@@ -46,15 +57,24 @@ class Zip
         $extracted = [];
         $location = $location ? $location : dirname($filename) . '/';
         $zip = new \ZipArchive();
+
         if ($zip->open($filename) === true) {
             $success = true;
             $numFiles = $zip->numFiles;
+
             for ($i = 0; $i < $numFiles && $success; $i++) {
                 $file = $zip->getNameIndex($i);
+
+                if ($file === false) {
+                    throw ZipException::nameOfIndexException();
+                }
+
                 $basename = basename($file);
+
                 if (preg_match($regex, $basename)) {
                     $basename = Uploads::validName(Strings::cleanName($basename), $location);
                     $copy = copy("zip://" . $filename . "#" . $file, $location . $basename);
+
                     if ($copy) {
                         $extracted[] = $basename;
                     } else {
@@ -62,40 +82,51 @@ class Zip
                     }
                 }
             }
+
             $zip->close();
+
             if (! $success) {
                 $countExtracted = count($extracted);
+
                 for ($i = 0; $i < $countExtracted; $i++) {
                     unlink($location . $extracted[$i]);
                 }
             }
+
             if ($success && $delete) {
                 unlink($filename);
             }
         }
+
         return $success ? $extracted : false;
     }
 
-    protected static function uniqueName($name, $folder)
+    protected static function uniqueName($name, $folder): string
     {
         if ($name) {
             $t = explode('.', $name);
-            if (count($t) == 1) {
+
+            if (count($t) === 1) {
                 $ext = '';
             } else {
-                $ext = "." . $t[count($t) - 1];
-                $t = array_slice($t, 0, count($t) - 1);
+                $ext  = '.' . $t[count($t) - 1];
+                $t    = array_slice($t, 0, count($t) - 1);
                 $name = implode('.', $t);
             }
-            $file = $name;
+
+            $file     = $name;
             $filepath = $folder . $name . $ext;
-            $i = 0;
+            $i        = 0;
+
             while (is_file($filepath)) {
                 $i++;
                 $filepath = $folder . $name . $i . $ext;
-                $file = $name . $i;
+                $file     = $name . $i;
             }
+
+            return $file . $ext;
         }
-        return $file . $ext;
+
+        throw ZipException::invalidNameException();
     }
 }
